Add missing XSRF cookie on import page
If this is the first form the user sees, they will not have any `_xsrf`
cookie set. JavaScript won't be able to read it and send it, causing a
403 error when trying to import a project.

Add the xsrf_form_html() to the form. While this is meant to create a
hidden form input that is not used here (because we don't submit the
form, and instead send a request from JavaScript), it will cause the
server to set the `_xsrf` cookie when the page is loaded.