Fix issue if document loads before it's added
If the POST to create the document completes before the /events, then we
would previously try to highlight the document in the list, but it
wouldn't be there yet. Handle this.

In addition, have updateDocumentsList() set the highlighting. This fixes
this issue, and additionally fixes cases where the highlighting would
disappear from the list, for example if deleting or renaming a document,
or if someone else adds a document.